Move plist_parser.py to a common place
`plist_parser.py` is used by the analyze, cmd and server part of the
CodeChecker too. To prepare source repository split up we move it to
a common place.
