+++ libcodechecker/plist_parser.py
