+++ tests/unit/test_plist_parser.py
@@ -29,7 +29,7 @@ from __future__ import absolute_import
 import os
 import unittest
 
-from libcodechecker.analyze import plist_parser
+from libcodechecker import plist_parser
 
 # These are the base skeletons for the main report sections where the
 # report hash and checker name is missing.
